requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pdfkit>=1.0.0
WeasyPrint>=57.0
PyYAML>=6.0 